+++ CHANGELOG.md
@@ -27,6 +27,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - expanded and improved the internal system roll formula generator
   - rolls with negative modifiers, that would result in a target value less than
     0, are now capped and don't break anymore
+- item images to the lists in the actor sheet
+- headings for the the lists in the inventory
 
 ### Changed